add button check
